requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
pytest>=6.2
pytest-django>=4.1
pytest-cov>=2.11
mixer>=7.1
djangorestframework>=3.12.2
